+++ werkzeug/exceptions.py
@@ -139,6 +139,19 @@ class HTTPException(Exception):
         response = self.get_response(environ)
         return response(environ, start_response)
 
+    def __str__(self):
+        return unicode(self).encode('utf-8')
+
+    def __unicode__(self):
+        if 'description' in self.__dict__:
+            txt = self.description
+        else:
+            txt = self.name
+        return '%d: %s' % (self.code, txt)
+
+    def __repr__(self):
+        return '<%s \'%s\'>' % (self.__class__.__name__, self)
+
 
 class _ProxyException(HTTPException):
     """An HTTP exception that expands renders a WSGI application on error."""